+++ library/src/com/cyphercove/flexbatch/batchable/Point2D.java
@@ -19,7 +19,6 @@ import com.badlogic.gdx.math.Vector2;
 import com.cyphercove.flexbatch.Batchable;
 import com.cyphercove.flexbatch.utils.AttributeOffsets;
 import com.cyphercove.flexbatch.utils.RenderContextAccumulator;
-import org.jetbrains.annotations.NotNull;
 
 /** A {@link Point} {@link Batchable Batchable} that supports a single texture at a time, with
  * two-dimensional position and color. It is designed to be drawn in a 2D plane.
@@ -45,7 +44,7 @@ public class Point2D extends Point<Point2D> {
 
 	/** Sets the position of the bottom left of the texture region in world space.
 	 * @return This object for chaining. */
-	public @NotNull Point2D position (float x, float y) {
+	public Point2D position (float x, float y) {
 		this.x = x;
 		this.y = y;
 		return this;
@@ -53,7 +52,7 @@ public class Point2D extends Point<Point2D> {
 
 	/** Sets the position of the bottom left of the texture region in world space.
 	 * @return This object for chaining. */
-	public @NotNull Point2D position (Vector2 position) {
+	public Point2D position (Vector2 position) {
 		x = position.x;
 		y = position.y;
 		return this;